- Moved iris regen to studio - Moved Iris LazyPregen into its own cmd - Fixed the mem leak "i hope" - Added Silent option to the lazy pregen
